Removed some 64-bit and other warnings.
git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@13258 32a0c1e6-9308-0410-80dd-f9d2257a62c7

+++ src/Libs/ri-util/eval.c
@@ -115,7 +115,7 @@ value EvalSizeofexpression(sizeofexpression soe)
 {
   type t = type_undefined;
   value v = value_undefined;
-  int i;
+  intptr_t i;
 
   if(sizeofexpression_expression_p(soe)) {
     expression e = sizeofexpression_expression(soe);
@@ -179,7 +179,7 @@ EvalIntrinsic(entity e, list la)
 value EvalConditionalOp(list la)
 {
   value vout, v1, v2, v3;
-  int arg1 = 0, arg2 = 0, arg3 = 0;
+  intptr_t arg1 = 0, arg2 = 0, arg3 = 0;
   bool failed = FALSE;
 
   pips_assert("Three arguments", gen_length(la)==3);
